refactor current code
Make the type constraint defined by a Macro, so that we don't have
to repeat the long "where" clause everywhere.

Add getBuffer() method to IStorage, because we have to get the concrete
buffer type when using CoopVec type.

Add a new associatedtype ComponentDataType to IVector, mainly because we
want a way to distinguish the IVector types, because we want the
InputVector and OutputVector to be the same type, so this is a good
pattern for us to add a new type constraint on the matmul method.